requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.12.2
    # via google
build==1.0.3
    # via pip-tools
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via pip-tools
duckdb==0.9.0
    # via -r requirements.in
google==3.0.0
    # via -r requirements.in
google-auth==2.23.2
    # via
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.1.0
    # via gspread
gspread==5.11.3
    # via -r requirements.in
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
numpy==1.26.0
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via
    #   build
    #   pytest
pandas==2.1.1
    # via -r requirements.in
pip-tools==7.3.0
    # via -r requirements.in
pluggy==1.3.0
    # via pytest
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pyproject-hooks==1.0.0
    # via build
pytest==7.4.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
requests==2.31.0
    # via requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
slack-sdk==3.22.0
    # via -r requirements.in
soupsieve==2.5
    # via beautifulsoup4
tabulate==0.9.0
    # via -r requirements.in
tzdata==2023.3
    # via pandas
urllib3==2.0.5
    # via requests
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
